made time interval smaller and changed time stuff to float, made player and agent speed and acc same as agents

+++ Assets/Scripts/Agent/Agent.cs
@@ -41,7 +41,7 @@ public class Agent : MonoBehaviour {
         GameObject newobs = Instantiate(blankObs,gameObject.transform);
         Observable obsInfo= newobs.GetComponent<Observable>();
 
-        int time = GameController.GetTime();
+        float time = GameController.GetTime();
         int zoneID = GameController.GetInstanceLevelController().GetZoneFromObj(gameObject);
 
         obsInfo.AddLocationClue(new LocationClue(agentId, zoneID, time));
